refactor seperate power class from powerSystem (aligns better with architecture - easier to test)
